+++ src/Cli.Tests/AddOpenTelemetryTests.cs
