+++ src/Cli.Tests/AddTelemetryTests.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-namespace Cli.Tests.TelemetryTests
+namespace Cli.Tests
 {
     /// <summary>
     /// Tests for verifying the functionality of adding telemetry to the config file.
@@ -10,7 +10,7 @@ namespace Cli.Tests.TelemetryTests
     public class AddTelemetryTests
         : VerifyBase
     {
-        public const string CONFIG_PATH = "test-telemetry-config.json";
+        public string RUNTIME_SECTION_WITH_APP_INSIGHTS_TELEMETRY_SECTION = GenerateRuntimeSection(TELEMETRY_SECTION_WITH_APP_INSIGHTS);
         public string RUNTIME_SECTION_WITH_EMPTY_TELEMETRY_SECTION = GenerateRuntimeSection(EMPTY_TELEMETRY_SECTION);
 
         [TestInitialize]
@@ -26,35 +26,42 @@ namespace Cli.Tests.TelemetryTests
         /// Testing to check telemetry options are correctly added to the config.
         /// Verifying scenarios such as enabling/disabling telemetry and providing a valid/empty connection string.
         /// </summary>
-        [TestMethod]
-        public void TestAddTelemetryBase(AddTelemetryOptions addTelemetryOptions, bool expectSuccess, out RuntimeConfig runtimeConfig)
+        [DataTestMethod]
+        [DataRow(CliBool.True, "", false, DisplayName = "Fail to add telemetry with empty app-insights connection string.")]
+        [DataRow(CliBool.True, "InstrumentationKey=00000000-0000-0000-0000-000000000000", true, DisplayName = "Successfully adds telemetry with valid connection string")]
+        [DataRow(CliBool.False, "InstrumentationKey=00000000-0000-0000-0000-000000000000", true, DisplayName = "Successfully adds telemetry but disabled")]
+        public void TestAddApplicationInsightsTelemetry(CliBool isTelemetryEnabled, string appInsightsConnString, bool expectSuccess)
         {
             MockFileSystem fileSystem = FileSystemUtils.ProvisionMockFileSystem();
-            fileSystem.AddFile(CONFIG_PATH, new MockFileData(INITIAL_CONFIG));
+            string configPath = "test-app-insights-config.json";
+            fileSystem.AddFile(configPath, new MockFileData(INITIAL_CONFIG));
 
             // Initial State
-            Assert.IsTrue(fileSystem.FileExists(CONFIG_PATH));
-            Assert.IsTrue(RuntimeConfigLoader.TryParseConfig(fileSystem.File.ReadAllText(CONFIG_PATH), out RuntimeConfig? config));
+            Assert.IsTrue(fileSystem.FileExists(configPath));
+            Assert.IsTrue(RuntimeConfigLoader.TryParseConfig(fileSystem.File.ReadAllText(configPath), out RuntimeConfig? config));
             Assert.IsNotNull(config);
             Assert.IsNotNull(config.Runtime);
             Assert.IsNull(config.Runtime.Telemetry);
 
             // Add Telemetry
             bool isSuccess = ConfigGenerator.TryAddTelemetry(
-                addTelemetryOptions,
+                new AddTelemetryOptions(appInsightsConnString: appInsightsConnString, appInsightsEnabled: isTelemetryEnabled, config: configPath),
                 new FileSystemRuntimeConfigLoader(fileSystem),
                 fileSystem);
 
-            runtimeConfig = config;
             // Assert after adding telemetry
             Assert.AreEqual(expectSuccess, isSuccess);
             if (expectSuccess)
             {
-                Assert.IsTrue(fileSystem.FileExists(CONFIG_PATH));
-                Assert.IsTrue(RuntimeConfigLoader.TryParseConfig(fileSystem.File.ReadAllText(CONFIG_PATH), out config));
+                Assert.IsTrue(fileSystem.FileExists(configPath));
+                Assert.IsTrue(RuntimeConfigLoader.TryParseConfig(fileSystem.File.ReadAllText(configPath), out config));
                 Assert.IsNotNull(config);
                 Assert.IsNotNull(config.Runtime);
                 Assert.IsNotNull(config.Runtime.Telemetry);
+                TelemetryOptions telemetryOptions = config.Runtime.Telemetry;
+                Assert.IsNotNull(telemetryOptions.ApplicationInsights);
+                Assert.AreEqual(isTelemetryEnabled is CliBool.True ? true : false, telemetryOptions.ApplicationInsights.Enabled);
+                Assert.AreEqual(appInsightsConnString, telemetryOptions.ApplicationInsights.ConnectionString);
             }
         }
 
@@ -66,41 +73,24 @@ namespace Cli.Tests.TelemetryTests
         [DataTestMethod]
         [DataRow(true, DisplayName = "Add AppInsights Telemetry when telemetry section is empty.")]
         [DataRow(false, DisplayName = "Overwrite AppInsights Telemetry when telemetry section already exists.")]
-        public void TestSetupAddTelemetryWhenTelemetryAlreadyExists(bool isEmptyTelemetry, string runtimeSectionWithTelemetry, TelemetryOption telemetryOption)
+        public void TestAddAppInsightsTelemetryWhenTelemetryAlreadyExists(bool isEmptyTelemetry)
         {
-            ArgumentNullException.ThrowIfNull(runtimeSectionWithTelemetry);
             MockFileSystem fileSystem = FileSystemUtils.ProvisionMockFileSystem();
-            string runtimeSection = isEmptyTelemetry ? RUNTIME_SECTION_WITH_EMPTY_TELEMETRY_SECTION : runtimeSectionWithTelemetry;
+            string configPath = "test-app-insights-config.json";
+            string runtimeSection = isEmptyTelemetry ? RUNTIME_SECTION_WITH_EMPTY_TELEMETRY_SECTION : RUNTIME_SECTION_WITH_APP_INSIGHTS_TELEMETRY_SECTION;
             string configData = $"{{{SAMPLE_SCHEMA_DATA_SOURCE},{runtimeSection}}}";
-            fileSystem.AddFile(CONFIG_PATH, new MockFileData(configData));
+            fileSystem.AddFile(configPath, new MockFileData(configData));
 
             // Initial State
-            Assert.IsTrue(fileSystem.FileExists(CONFIG_PATH));
-            Assert.IsTrue(RuntimeConfigLoader.TryParseConfig(fileSystem.File.ReadAllText(CONFIG_PATH), out RuntimeConfig? config));
+            Assert.IsTrue(fileSystem.FileExists(configPath));
+            Assert.IsTrue(RuntimeConfigLoader.TryParseConfig(fileSystem.File.ReadAllText(configPath), out RuntimeConfig? config));
             Assert.IsNotNull(config);
             Assert.IsNotNull(config.Runtime);
             Assert.IsNotNull(config.Runtime.Telemetry);
 
             if (isEmptyTelemetry)
             {
-                switch (telemetryOption)
-                {
-                    case TelemetryOption.ApplicationInsights:
-                        Assert.IsNull(config.Runtime.Telemetry.ApplicationInsights);
-                        break;
-
-                    case TelemetryOption.OpenTelemetry:
-                        Assert.IsNull(config.Runtime.Telemetry.OpenTelemetry);
-                        break;
-
-                    case TelemetryOption.AzureLogAnalytics:
-                        Assert.IsNull(config.Runtime.Telemetry.AzureLogAnalytics);
-                        break;
-
-                    default:
-                        Assert.Fail("Invalid Option was chosen");
-                        break;
-                }
+                Assert.IsNull(config.Runtime.Telemetry.ApplicationInsights);
             }
             else
             {
@@ -111,14 +101,14 @@ namespace Cli.Tests.TelemetryTests
 
             // Add Telemetry
             bool isSuccess = ConfigGenerator.TryAddTelemetry(
-                new AddTelemetryOptions(appInsightsConnString: "InstrumentationKey=11111-1111-111-11-1", appInsightsEnabled: CliBool.False, config: CONFIG_PATH),
+                new AddTelemetryOptions(appInsightsConnString: "InstrumentationKey=11111-1111-111-11-1", appInsightsEnabled: CliBool.False, config: configPath),
                 new FileSystemRuntimeConfigLoader(fileSystem),
                 fileSystem);
 
             // Assert after adding telemetry
             Assert.IsTrue(isSuccess);
-            Assert.IsTrue(fileSystem.FileExists(CONFIG_PATH));
-            Assert.IsTrue(RuntimeConfigLoader.TryParseConfig(fileSystem.File.ReadAllText(CONFIG_PATH), out config));
+            Assert.IsTrue(fileSystem.FileExists(configPath));
+            Assert.IsTrue(RuntimeConfigLoader.TryParseConfig(fileSystem.File.ReadAllText(configPath), out config));
             Assert.IsNotNull(config);
             Assert.IsNotNull(config.Runtime);
             Assert.IsNotNull(config.Runtime.Telemetry);
@@ -130,7 +120,7 @@ namespace Cli.Tests.TelemetryTests
         /// <summary>
         /// Generates a JSON string representing a runtime section of the config, with a customizable telemetry section.
         /// </summary>
-        protected static string GenerateRuntimeSection(string telemetrySection)
+        private static string GenerateRuntimeSection(string telemetrySection)
         {
             return $@"
                 ""runtime"": {{
@@ -159,16 +149,21 @@ namespace Cli.Tests.TelemetryTests
         }
 
         /// <summary>
+        /// Represents a JSON string for the telemetry section of the config, with Application Insights enabled and a specified connection string.
+        /// </summary>
+        private const string TELEMETRY_SECTION_WITH_APP_INSIGHTS = @"
+            ""telemetry"": {
+                ""application-insights"": {
+                    ""enabled"": true,
+                    ""connection-string"": ""InstrumentationKey=00000000-0000-0000-0000-000000000000""
+                }
+            }";
+
+        /// <summary>
         /// Represents a JSON string for the empty telemetry section of the config.
         /// </summary>
         private const string EMPTY_TELEMETRY_SECTION = @"
             ""telemetry"": {}";
-
-        enum TelemetryOption
-        {
-            ApplicationInsights,
-            OpenTelemetry,
-            AzureLogAnalytics
-        }
     }
+
 }
